+++ examples/generation_gpt_neo.rs
@@ -17,7 +17,7 @@ extern crate anyhow;
 use rust_bert::gpt_neo::{
     GptNeoConfigResources, GptNeoMergesResources, GptNeoModelResources, GptNeoVocabResources,
 };
-use rust_bert::pipelines::common::ModelType;
+use rust_bert::pipelines::common::{ModelResources, ModelType};
 use rust_bert::pipelines::text_generation::{TextGenerationConfig, TextGenerationModel};
 use rust_bert::resources::RemoteResource;
 use tch::Device;
@@ -38,7 +38,7 @@ fn main() -> anyhow::Result<()> {
     ));
     let generate_config = TextGenerationConfig {
         model_type: ModelType::GPTNeo,
-        model_resource,
+        model_resource: ModelResources::TORCH(model_resource),
         config_resource,
         vocab_resource,
         merges_resource: Some(merges_resource),